+++ game/server/sources/units/Unit.cpp
@@ -2,7 +2,7 @@
 #include <utility>
 #include "server/headers/units/Unit.h"
 
-Unit::Unit(int id, double hp, int range, int speed, int type, coordenada_t coord) :
+Unit::Unit(int id, unsigned int hp, unsigned int range, unsigned int speed, int type, coordenada_t coord) :
         Entity(id, hp, type, coord), range(range), speed(speed) {}
 
 void Unit::setPath(std::stack<coordenada_t> _path) {
